changed database migration to bigint

+++ db/migrate/20130416234711_create_friend_pages.rb
@@ -2,7 +2,7 @@ class CreateFriendPages < ActiveRecord::Migration
   def change
     create_table :friend_pages do |t|
       t.integer :user_id
-      t.integer :fb_id
+      t.bigint :fb_id
       t.string :location
       t.text :notes
